+++ NEWS.md
@@ -1,3 +1,26 @@
+# manynet 0.2.7
+
+2023-11-01
+
+## Making
+
+- Updated treatment of adjacency matrices in `as_igraph()` in accordance with upcoming updates to `{igraph}` package (closing #27)
+
+## Manipulating
+
+- Added `to_mentoring` function
+
+## Mapping
+
+- Fixed bugs in `autographr()` related to `edge_size` and `edge_color`
+- Fixed bugs and updated label placements for "circular" and "concentric" layouts for `autographr()`
+- Fixed issues with self-ties in `autographr()`
+- Updated tests for `autographr()`
+
+## Data
+
+- Added `ison_friends`, a one-mode network on character connections of a popular TV series
+
 # manynet 0.2.6
 
 2023-10-25